Add shared ButtonLayout component + DualSense BT/mic fixes (D1, D3)
Button Layout Component (D3):
  - src/ui/button_layout.h: Shared ControllerState-driven button layout
    with 20 buttons, analog sticks, triggers, touchpad, gyro/accel.
  - src/ui/button_layout.cpp: ImGui renderer with PlayStation button
    colors, stick position overlays, trigger bars, finger tracking,
    connection status, 'tested' checklist.

DualSense Bluetooth fix (D1):
  - BT report[2] changed from 0x10 to 0x00 to avoid phantom audio
    enable that broke BT mic/speaker functionality.
  - Mic audio data extraction: 4-channel × 8-sample PCM16 from input
    report offset o+74, gated by status bit 0x04.

PENDING.md: Updated with Phase D (DualSense fixes) and Phase R (VRR,
FSR, HDR) — fully populated with actionable sub-tasks.

+++ src/gpu/dualsense_hid.h
@@ -427,7 +427,7 @@ namespace DualSense {
                     s.mic_channels = 4;
                     for (int ch = 0; ch < 4; ++ch) {
                         for (int smp = 0; smp < 8; ++smp) {
-                            const int off = o + 74 + (ch * 16) + (smp * 2);
+                            const size_t off = o + 74 + static_cast<size_t>(ch * 16 + smp * 2);
                             if (off + 2 <= static_cast<int>(len)) {
                                 s.mic_samples[ch][smp] = static_cast<s16>(
                                     static_cast<s16>(r[off]) |
